+++ plugin_info.toml
@@ -13,7 +13,7 @@ license = 'MIT'
 
 [plugin-install]
 #packages required for your plugin:
-packages-required = ['pipython', 'bitstring', 'pymodaq>=4.1.0']
+packages-required = ['pipython', 'bitstring', 'pymodaq>=4.1.0', 'msl-loadlib']
 
 [features]  # defines the plugin features contained into this plugin
 instruments = true  # true if plugin contains instrument classes (else false, notice the lowercase for toml files)
